+++ fast_curator/write.py
@@ -14,7 +14,7 @@ __all__ = ["known_expanders", "prepare_file_list", "write_yaml",
 
 
 def prepare_file_list(files, dataset, eventtype, tree_name, expand_files="xrootd",
-                      absolute_paths=True, no_empty_files=True):
+                      absolute_paths=True, no_empty_files=True, confirm_tree=True):
     """
     Expands all globs in the file lists and creates a dataframe similar to those from a DAS query
     """
@@ -24,7 +24,9 @@ def prepare_file_list(files, dataset, eventtype, tree_name, expand_files="xrootd
     full_list = expand_files.expand_file_list(files)
     if absolute_paths:
         full_list = [os.path.realpath(f) if ':' not in f else f for f in full_list]
-    full_list, numentries = expand_files.check_entries(full_list, tree_name, no_empty=no_empty_files)
+    full_list, numentries = expand_files.check_entries(full_list, tree_name,
+                                                       no_empty=no_empty_files,
+                                                       confirm_tree=confirm_tree)
 
     data = {}
     data["eventtype"] = eventtype